+++ packages/plugin-packages/mashroom-portal-sandbox-app/src/js/components/MessageBusSendForm.tsx
@@ -4,11 +4,12 @@ import {
     Form,
     SelectField,
     SourceCodeEditorField,
-    Button
+    Button,
+    TextareaField
 } from '@mashroom/mashroom-portal-ui-commons';
 
 import type {ReactNode} from 'react';
-import type {MashroomPortalMessageBus} from '@mashroom/mashroom-portal/type-definitions';
+import type {MashroomPortalMessageBus, MashroomPortalStateService} from '@mashroom/mashroom-portal/type-definitions';
 import type {FormContext} from '@mashroom/mashroom-portal-ui-commons/type-definitions';
 import type {ActivePortalApp, MessageBusMessage} from '../types';
 
@@ -19,9 +20,11 @@ type FormData = {
 
 type Props = {
     messageBus: MashroomPortalMessageBus;
+    portalStateService: MashroomPortalStateService;
     activePortalApp: ActivePortalApp | undefined | null;
     topicsSubscribedByApp: Array<string>;
     addMessagePublishedBySandbox: (messageBus: MessageBusMessage) => void;
+    sbAutoTest: boolean;
 }
 
 export default class MessageBusSendForm extends PureComponent<Props> {
@@ -54,7 +57,9 @@ export default class MessageBusSendForm extends PureComponent<Props> {
     onSubmit(values: FormData, context: FormContext): void {
         const { messageBus, addMessagePublishedBySandbox } = this.props;
         const { topic, message } = values;
+
         const data = JSON.parse(message);
+        
         messageBus.publish(topic, data);
         addMessagePublishedBySandbox({
             topic,
@@ -81,7 +86,9 @@ export default class MessageBusSendForm extends PureComponent<Props> {
                         <SelectField id='mashroom-sandbox-publish-message-topic' name='topic' labelId='topic' options={topicOptions} emptyOption={true} />
                     </div>
                     <div className='mashroom-sandbox-app-form-row'>
-                        <SourceCodeEditorField id='mashroom-sandbox-publish-message-message' name="message" labelId='message' language='json' theme='light' height={120} />
+                        { !this.props.sbAutoTest ?
+                        <SourceCodeEditorField id='mashroom-sandbox-publish-message-message' name="message" labelId='message' language='json' theme='light' height={120} /> : 
+                        <TextareaField id='mashroom-sandbox-publish-message-message' name="message" labelId='message' /> }
                     </div>
                     <div className='mashroom-sandbox-app-form-button-row'>
                         <Button id='mashroom-sandbox-publish-message' type='submit' labelId='sendMessage'/>