+++ packages/plugin-packages/mashroom-portal-sandbox-app/src/js/types.ts
@@ -66,4 +66,5 @@ export type PortalAppQueryParams = {
     lang: string | undefined | null;
     permissions: any | undefined | null;
     appConfig: any | undefined | null;
+    autoTest: boolean;
 }